+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com) and this project adheres to [Semantic Versioning](http://semver.org).
 
+## UNRELEASED
+
+### Fixed
+
+- Fixed bug in client analytics event log store that caused events not to be ordered correctly, breaking sequence number recording
+
 ## 0.9.0 - 2018-09-16
 
 ### Added